Refactor: create directory "parsers" and move all parser modules into it.
